+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Improvement:
 * [OLMIS-3306](https://openlmis.atlassian.net/browse/OLMIS-3306): Added role type column on the role list screen.
 * [OLMIS-2285](https://openlmis.atlassian.net/browse/OLMIS-2285): Supported program start date is now editable on facility edit screen.
+* [OLMIS-3403](https://openlmis.atlassian.net/browse/OLMIS-3403): Added query params for get facilities.
 
 5.3.0 / 2017-11-09
 ==================